# Conflicts:
#	.idea/deploymentTargetSelector.xml

+++ app/src/main/java/com/appmoviles/utadeliciasapp/Products.kt
@@ -7,5 +7,5 @@ data class Products(
     val descripcion: String = "",
     val imagen: String = "",
     val cantidad: Int = 0,
-    val precio: Int = 0
+    val precio: Double = 0.0 // Cambiar a Double
 )
